% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmontAval.R
\name{AmontAval}
\alias{AmontAval}
\title{Buffer dissymetrique}
\usage{
AmontAval(lig, r, amont = 50, aval = 150, pas = 25, conc = 1.8)
}
\arguments{
\item{lig}{= objet de type LINESTRING correspondant à une piste ou une route.}

\item{r}{= raster correspondant à un MNT.}

\item{amont}{= distance selon la pente exploitable en amont. Par défaut amont = 50m.}

\item{aval}{= distance selon la pente exploitable en aval. Par défaut aval = 150m.}

\item{pas}{= résolution.  Par défaut pas = 25m}

\item{conc}{= indice relatif de concavité. Plus cet indice est faible plus le contour sera découpé.
Cet indice ne doit pas être inférieur à 1. Par défaut cet indice est égal à 1.8.}
}
\description{
Cette fonction est nécessaire pour cartographier les zones exploitables par skidder en montagne.
Elle renvoie un polygone correspondant à un buffer dissymétrique amont aval autour d'un seul LINESTRING.
}
\examples{

}
\author{
Bruciamacchie Max
}
