% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BufferDis.R
\name{BufferDis}
\alias{BufferDis}
\title{Buffer dissymetrique}
\usage{
BufferDis(line, r, amont = 50, aval = 150, pas = 50)
}
\arguments{
\item{line}{= objet de type LINE correspondant à une piste ou une route.}

\item{r}{= raster correspondant à un MNT.}

\item{amont}{= distance selon la pente exploitable en amont. Par défaut amont = 50m.}

\item{aval}{= distance selon la pente exploitable en aval. Par défaut aval = 150m.}

\item{pas}{= résolution.  Par défaut pas = 25m}
}
\description{
Cette fonction renvoie un polygone correspondant à un buffer dissymétrique amont aval.
Elle permet de cartographier les zones exploitables par skidder en montagne.
}
\examples{

}
\author{
Bruciamacchie Max
}
