% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BufferDissy.R
\name{BufferDissy}
\alias{BufferDissy}
\title{Buffer dissymetrique}
\usage{
BufferDissy(shp, r, amont = 50, aval = 150, pas = 50, conc = 1.8)
}
\arguments{
\item{r}{= raster correspondant à un MNT.}

\item{amont}{= distance selon la pente exploitable en amont. Par défaut amont = 50m.}

\item{aval}{= distance selon la pente exploitable en aval. Par défaut aval = 150m.}

\item{pas}{= résolution.  Par défaut pas = 25m}

\item{conc}{= indice relatif de concavité. Plus cet indice est faible plus le contour sera découpé.
Cet indice ne doit pas être inférieur à 1. Par défaut cet indice est égal à 1.8.}

\item{line}{= objet de type LINE correspondant à une piste ou une route.}
}
\description{
Cette fonction renvoie un polygone correspondant à un buffer dissymétrique amont aval.
Elle permet de cartographier les zones exploitables par skidder en montagne.
}
\examples{

}
\author{
Bruciamacchie Max
}
