% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CadCorFor.R
\name{CadCorFor}
\alias{CadCorFor}
\title{Correspondance parcellaire forestier et cadastral}
\usage{
CadCorFor(shpCad, shpFor, nom = "ParcellaireForestier")
}
\arguments{
\item{shpCad}{= fichier du parcellaire cadastral au format sf.}

\item{shpFor}{= fichier du parcellaire forestier au format sf.}

\item{Nom}{= Nom par défaut du fichier de sortie.}
}
\value{
La fonction

- renvoie un tableau de correspondance entre numéros de parcelle cadastrale et numéro de parcelle forestiére.

- renvoie un tableau des surfaces des parcelles forestières

- enregistre le shape des parcelles forestières avec comme nouveau champ, un coefficient permettant de passer de la surface
SIG à la surface cadastrale;
}
\description{
La fonction recherche la correspondance entre parcellaire forestier et cadastral.
Elle fait en sorte que la somme des surfaces des parcelles forestières soit égale à la surface
officielle cadastrale. Cette recherche est faite en évitant de trop aggréger les parcelles.
Elle calcule un coefficient par parcelle permettant de passer de la surface SIG à la surface cadastrale.

La fonction nécessite en entrée 2 fichiers géoréférencés au format sf : parcellaire cadastral et forestier.
Si les fichiers en entrée sont au format shp, l'exemple ci-après montre comment les transformer au format sf.

La table attributaire du parcellaire cadastral doit contenir au moins les champs suivants :
Commune, Section, Num et Surface.
}
\examples{
library(sf, quietly =T, warn.conflicts =F)
library(ForestTools)
##### Données
data("parCadast")
data("parFor")
parCadast <- st_as_sf(parCadast)
parFor <- st_as_sf(parFor)
##### Visualisation
par(mar=c(0,0,0,0))
plot(st_geometry(parFor), lwd=2.5, border='blue')
plot(st_geometry(parCadast), add=TRUE, border='red')
##### Utilisation fonction
res <- CadCorFor(parCadast, parFor)
##### Résultats
res$Correspondance
res$Surfaces

}
\author{
Bruciamacchie Max
}
