% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CadHtml2csv.R
\name{CadHtml2csv}
\alias{CadHtml2csv}
\title{Lecture matrices cadastrales .html}
\usage{
CadHtml2csv(enrg = FALSE)
}
\arguments{
\item{enrg}{= si ce paramètre est TRUE, le tableau est enregistré au format .csv}
}
\value{
La fonction retourne un data frame et permet de l'enregistrer au format csv.
}
\description{
Lecture de matrices cadastrales au format html contenues dans un dossier,
construction d'un tableau contenant le nom de la commune, la section, le numéro de parcelle,
la nature d'occupation du sol et la surface.
}
\examples{
\donttest{
####### Utilisation
tab <- CadHtml2csv(enrg=TRUE)
}

}
\author{
Bruciamacchie Max & Chevereau Matthieu
}
