% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CadPar.R
\name{CadPar}
\alias{CadPar}
\title{Selection de parcelles cadastrales}
\usage{
CadPar(df, shp)
}
\arguments{
\item{df}{= tableau contenant la liste des parcelles cadastrales. Il doit contenir une colonne nommée "idu" ou "IDU"}

\item{shp}{= un fichier géoréférencé contenant des parcelles cadastrales.
Sa table attributaire doit contenir une colonne nommée "idu" ou "IDU"}
}
\value{
La fonction retourne un shapefile polygone au format shp.
}
\description{
La fonction permet de produire un shapefile polygone contenant l'ensemble des parcelles cadastrales
d'un ou plusieurs propriétaires. La fonction sélectionne dans un fichier .shp les parcelles cadastrales
figurant dans des extraits de matrice cadastrale.
}
\examples{
\donttest{
####### Utilisation
data("MatriceCad")
data("Parcellaire")
shp <- CadPar(MatriceCad, Parcellaire)
plot(st_geometry(Parcellaire), border='blue', lwd=0.5)
plot(st_geometry(shp), border='red', add=T)
}

}
\author{
Bruciamacchie Max & Chevereau Matthieu
}
