% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimatOmbro.R
\name{ClimatOmbro}
\alias{ClimatOmbro}
\title{Diagramme ombrothermique}
\usage{
ClimatOmbro(temp, precip)
}
\arguments{
\item{temp}{= vecteur contenant les températures}

\item{precip}{= vecteur contenant les précipitations}
}
\value{
La fonction renvoie graphique au format ggplot.
}
\description{
Edition d'un diagramme ombrothermique à partir d'informations mensuelles sur les températures
et les précipitations.
}
\examples{
Precipitation = c(92,82,70,64,49,38,29,31,46,68,96,106)
Temperature = c(6.4,7.6,9.6,11.6,15.4,18.3,20.8,20.9,18.1,14.2,9.4,7.3)
ClimatOmbro(Precipitation, Temperature)

}
\author{
Bruciamacchie Max et étudiants FIF
}
