% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteo.R
\name{ClimatOmbroMF}
\alias{ClimatOmbroMF}
\title{Diagramme ombrothermique données météoFrance}
\usage{
ClimatOmbroMF(df)
}
\arguments{
\item{df}{= tableau contenant les informations météo}
}
\value{
La fonction renvoie graphique au format ggplot.
}
\description{
Edition d'un diagramme ombrothermique à partir d'un tableau contenant les informations
classiques de météoFrance : Station, Année, Mois, RR (Précipitations), TN (témpérature minimale) et
TX (température maximale). Si le tableau en entrée contient plusieurs stations, une boîte de dialogue
permet d'en sélectionner une seule.
}
\examples{
##### Données
data("DataMeteo")
##### Visualisation
head(DataMeteo)
##### Utilisation fonction
ClimatOmbroMF(DataMeteo)

}
\author{
Bruciamacchie Max et étudiants FIF
}
