% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contour2Line.R
\name{Contour2Line}
\alias{Contour2Line}
\title{Transformation raster en courbes de niveau}
\usage{
Contour2Line(shp, r, nlevels = 40, valline = NULL)
}
\arguments{
\item{shp}{= périmètre de la zone d'étude au format sf.}

\item{r}{= raster à convertir en courbe.}

\item{nlevels}{= nombre de courbes. Par défaut nlevels = 40.}

\item{valline}{= valeur de la ligne à extraire}
}
\value{
La fonction renvoie un objet LINESTRING.
}
\description{
La fonction transforme un modèle numérique de terrain en courbes de niveau.
puis offre la possibilité de sélectionner une de ces courbes de niveau comme limite de
séparation d'un polygone.
}
\examples{
library(sf, quietly =T, warn.conflicts =F)
library(Forestree)
data(razel)
mnt = SiteMnt(razel)
courbe <- Contour2Line(razel, mnt, valline="430")

plot(st_geometry(razel))
plot(st_geometry(courbes), add=T, col='orange', lwd=0.5)
plot(st_geometry(Ligne), add=T, col='red')

}
\author{
Bruciamacchie Max
}
