% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CourseSoleil.R
\name{CourseSoleil}
\alias{CourseSoleil}
\title{Course du soleil}
\usage{
CourseSoleil(Lat, Lon)
}
\arguments{
\item{Lat}{= latitude.}

\item{Lon}{= longitude.}
}
\description{
Cette fonction renvoie la hauteur du soleil exprimée en degré au premier jour du printemps
(20 mars) et de l'été (20 juin).
}
\examples{
\donttest{
library(tidyverse)
library(suncalc)

res <- CourseSoleil(Lat = 50.1, Lon = 1.83)
res$graph
}
}
\author{
Bruciamacchie Max
}
