% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecoupRN.R
\name{DecoupRN}
\alias{DecoupRN}
\title{Régions naturelles}
\usage{
DecoupRN(formatSf = TRUE, dept = TRUE)
}
\arguments{
\item{formatSf}{= choix du format. Si formatSf = TRUE l'objet sera au format sf, sinon il sera au format shp.
Par défaut formatSf = TRUE.}

\item{dept}{= échelle du découpage. Si dept = TRUE, la France sera découpée par département. Par défaut dept = TRUE}
}
\value{
La fonction renvoie un objet géoréférencé de la France au format shp ou sf.
}
\description{
La fonction met à disposition le découpage de la France en régions naturelles détaillées ou non
par département, ainsi que le découpage en sylvoécorégions.
}
\examples{
 rn <- DecoupRN(formatSf = T, dept = F)
 class(rn)
 rn <- DecoupRN(formatSf = F, dept = F)
 class(rn)

}
\author{
Bruciamacchie Max
}
