% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DessertePL.R
\name{DessertePL}
\alias{DessertePL}
\title{Profil en long}
\usage{
DessertePL(shp, r, pas = 25)
}
\arguments{
\item{shp}{= objet sf de type LINESTRING correspondant à une piste ou une route.}

\item{r}{= raster correspondant à un MNT.}

\item{pas}{= résolution.  Par défaut pas = 25m}
}
\description{
Cette fonction permet de visualiser le profil en long de tout éléments de la desserte.
}
\examples{

}
\author{
Bruciamacchie Max
}
