% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoInd.R
\name{EcoTIR}
\alias{EcoTIR}
\title{Calcul du Taux Interne de Rentabilite}
\usage{
EcoTIR(age, cash)
}
\arguments{
\item{age}{= vecteur d'années d'apparition des flux financiers.}

\item{Montant}{= vecteur des montants annuels des différents flux financiers. Il s'agit de montants nets
(recettes - dépenses). Ce vecteur doit forcément contenir des valeurs négatives et positives.}
}
\description{
Cette fonction calcule le taux interne de rentabilite d'une série de cashflows.
}
\examples{
df <- data.frame(Année = c(2, 4, 6, 30),
                 Montant = c(-0.6, -0.9, -1.5, 9))
EcoTIR(df$Montant, df$Année)

}
\author{
Bruciamacchie Max
}
