% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvolEssence.R
\name{EvolEssence}
\alias{EvolEssence}
\title{Evolution de la répartition du volume d'une essence.}
\usage{
EvolEssence(ess = "09", group = "greco")
}
\arguments{
\item{ess}{= code d'une ou plusieurs essences selon la codification de l'IFN}

\item{group}{= l'argument group n'admet que les modalités greco (11 grandes régions écologiques)
ou climat (décomposition du territoire national en 8 climats). Par défaut l'argument group pend la valeur
"greco".}
}
\description{
Evolution de la répartition du volume d'une essence par classe de diamètre et par
région.
}
\examples{
data(IFNarbres)
data(IFNplacettes)
data(ser)
EvolEssence(ess="05", group="climat")

}
\author{
BRUCIAMACCHIE Max
}
