% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploiTracteur.R, R/Zone Exploitable.R
\name{ExploiTracteur}
\alias{ExploiTracteur}
\title{Zones exploitables}
\usage{
ExploiTracteur(r, shp)

ExploiTracteur(r, shp)
}
\arguments{
\item{r}{= un modèle numérique de terrain}

\item{shp}{= la route au format sf}
}
\value{
La fonction renvoit le MNT

La fonction renvoit le MNT
}
\description{
Affichage des zones exploitables (pente<0.30) par tracteur sur le MNT

Affichage des zones exploitables (pente<0.30) par tracteur sur le MNT
}
\author{
Max Bruciamacchie

Max Bruciamacchie
}
