% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmoniseRaster.R
\name{HarmoniseRaster}
\alias{HarmoniseRaster}
\title{Harminisation de 2 rasters}
\usage{
HarmoniseRaster(r1, r2, resout = 1)
}
\arguments{
\item{r1}{= RasterLayer qui sert de référence}

\item{r2}{= RasterLayer à harmoniser}

\item{resout}{= résolution des deux rasters en sortie. Par défaut resout = 1m.}
}
\value{
La fonction renvoie un stack contenant les deux rasters.
}
\description{
La fonction harmonise le deuxième raster de sorte qu'il ait même résolution, origine
et extension que le premier.
}
\examples{
file <- system.file("MNS.tif", package = "Forestree")
mns <- raster(file)
file <- system.file("MNT.tif", package = "Forestree")
mnt <- raster(file)
result <- HarmoniseRaster(mns, mnt, resout = 1)
}
\author{
Bruciamacchie Max
}
