% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNdata.R
\encoding{UTF-8}
\name{IFNdata}
\alias{IFNdata}
\title{Rechargement des donnees IFN}
\usage{
IFNdata(debut, fin)
}
\arguments{
\item{debut}{= Première année pour laquelle on souhaite obtenir des données.
Attention, l'année de début doit être antérieure à 2008}

\item{fin}{= Dernière année pour laquelle on souhaite obtenir des données}
}
\value{
Cette fonction enregistre dans une archive IFNdata.Rdata deux tables : IFNarbres et IFNplacettes.
}
\description{
Cette fonction permet de télécharger les données brutes de l'IFN à partir du site
http://inventaire-forestier.ign.fr/spip/IMG/zip/.
}
\examples{
IFNdata(2011,2016)

}
\author{
HIGUET Valentin, BRUCIAMACCHIE Max
}
