% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNdendro.R
\name{IFNdendro}
\alias{IFNdendro}
\title{Informations dendrométriques déduites de l'IFN}
\usage{
IFNdendro(shp = NULL, rep = NULL, TailleBuffer = 450, enreg = F)
}
\arguments{
\item{shp}{= fichier géoréférencé sous forme de polygone(s).}

\item{TailleBuffer}{= Taille du buffer en mètre. Par défaut elle est égale à 450 m.}

\item{enreg}{= argument permettant de choisir l'enregistrement ou pas du tableau au format .csv.
Par défaut enreg=FALSE.}
}
\value{
La fonction renvoie un tableau des surfaces terrières : totale, par essence, par structure et
pour les BM + GB des 2 essences les plus représentées.
}
\description{
Calcul de la surface terrière des principales essences, de la structure en PER, PB, BM, GB et de
l'importance en BM + GB des deux essences contribuant le plus à la surface terrière totale. Toutes ces informations sont
accompagnées de leur coefficient de variation. Les calculs ont été faits en supposant que toutes les placettes
ont le même poids.
La fonction nécessite en entrée un shape correspondant au périmètre retenu : forêt, massif, sylvoécorégion, etc.
}
\examples{
library(sf)
data(perim)
load("IFNdata.Rdata")
res <- IFNdendro(shp=perim)
res$tab        # Répartition de la surface terrière
res$Nb         # nombre de placettes IFN
res$perimetre  # périmètre retenu
res$placettes  # placettes IFN
plot(st_geometry(perim))
plot(st_geometry(res$placettes), add=T, pch=3, col='red')

}
\author{
Bruciamacchie Max
}
