% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JP2toTIF.R
\name{JP2toTIF}
\alias{JP2toTIF}
\title{Conversion JP2 to TIF}
\usage{
JP2toTIF()
}
\description{
Cette fonction convertit tous les fichiers au format JP2 contenus dans un dossier en format <<tif.
}
\examples{
JP2toTIF()

}
\author{
Bruciamacchie Max
}
