% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosaic.R
\name{Mosaic}
\alias{Mosaic}
\title{Assembler rasters}
\usage{
Mosaic(Nom)
}
\description{
Cette fonction assemble tous les rasters au format TIF ou ASC contenus dans un 
dossier en un seul fichier au format TIF.
}
\examples{
Mosaic(Nom = "IRC.tif")

}
\author{
Bruciamacchie Max
}
