% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProtectData.R
\name{NetProtectData}
\alias{NetProtectData}
\title{Recherche statuts de protection}
\usage{
NetProtectData(shp, width = 50, inter = TRUE, onedir = FALSE)
}
\arguments{
\item{shp}{= périmètre du territoire au format sf ou shp.}

\item{width}{= taille du buffer en mètre. Par défaut width= 50 m.}

\item{inter}{= Par défaut il n'y a pas intersection avec le perimètre (inter=FALSE)}

\item{onedir}{= Par défaut les différents statuts sont rangés dans des dossiers différents
(onedir=FALSE). Si onedir = TRUE, les différents statuts sont rangés dans un seul dossier
appelé StatutProtect.}
}
\value{
La fonction crée un dossier appelé StatutProtect qui contiendra tous les statuts de protection
existant sur la propriété.
}
\description{
La fonction recherche sur internet les statuts de protection applicables sur un territoire donné.
La fonction nécessite en entrée un fichier géoréférencé du périmètre au format sf ou shp.
Si les fichiers en entrée sont au format shp la fonction les transforme au format sf.
}
\examples{
library(sf, quietly =T, warn.conflicts =F)
library(tools, quietly =T, warn.conflicts =F)
##### Données
data(perim)
NetProtectData(perim, width=100)

}
\author{
Bruciamacchie Max
}
