% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rayonnement.R
\name{Rayonnement}
\alias{Rayonnement}
\title{Cartographie du rayonnement maximal sur un territoire}
\usage{
Rayonnement(shp, dem, buffer = 100)
}
\arguments{
\item{shp}{= périmètre de la portion de territoire.}

\item{dem}{= raster correspondant au MNT du périmètre.}

\item{buffer}{= buffer à appliquer au périmètre. Par défaut buffer=100 m.}
}
\description{
Cette fonction nécessite en entrée le périmètre du territoire sous forme d'objet sf
(library sf) ainsi que son MNT. Elle renvoie un raster des écarts de rayonnement.
}
\examples{
\donttest{
library(sf)
library(tidyverse)
library(raster)
library(insol)

data(razel)
r <- SiteMnt(razel)
rad <- Rayonnement(razel, r)

plot(rad, axes=F, box=F)
plot(st_geometry(razel), add=T)
# writeRaster(rad, "Rad.tif", format="GTiff", overwrite=TRUE)
}
}
\author{
Bruciamacchie Max
}
