% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteTopo.R
\name{SiteTopo}
\alias{SiteTopo}
\title{Indicateurs topographiques}
\usage{
SiteTopo(shp, r)
}
\arguments{
\item{shp}{= périmètre de la zone au format sf.}

\item{r}{= MNT sous forme de raster.}
}
\value{
La fonction renvoie trois indicateurs topographiques : pente, classes de pente par exposition,
direction de la ligne de plus grande pente.
}
\description{
La fonction permet de calculer des indicateurs topographiques sur un périmètre donné.
}
\examples{
\donttest{
library(raster)
library(sf)
library(tidyverse)
library(scales)
res <- SiteTopo(perim, mnt)
res$Pente
res$ExpoPente
res$Ecoul
}

}
\author{
Bruciamacchie Max
}
