% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnionPolyDeleteHole.R
\name{UnionPolyDeleteHole}
\alias{UnionPolyDeleteHole}
\title{Simplification du domaine d'étude - union et suppression des trous}
\usage{
UnionPolyDeleteHole(nom = "PerimCor")
}
\arguments{
\item{nom}{= nom souhaité pour le fichier en sortie.}
}
\value{
La fonction renvoie le shape PerimCor.shp comme polygone unifié.
}
\description{
La fonction UnionPolyDeleteHole utilise en entrée un shape de type surfacique.
Elle agrège tous les polygones en un seul, et supprime les éventuels trous.\cr
}
\author{
Bruciamacchie Max
}
