% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValeurVenale.R
\name{ValeurVenale}
\alias{ValeurVenale}
\title{Valeur des forêts}
\usage{
ValeurVenale(an, minSurf = 1, maxPrix = 50000)
}
\arguments{
\item{an}{= année. Ce paramètre n'accepte actuellement que les années 2016, 2017 et 2018.}

\item{minSurf}{= seuil minimal de surface en ha. Par défault minSurf = 1 ha.}

\item{maxPrix}{= seuil maximal de prix à l'hectare. Par défault maxPrix = 50000 euro/ha.}
}
\description{
Les valeurs des forêts sont issues de la base  DVF (Demandes de valeurs foncières)
du site data.gouv.fr. Les deux département de l'Alsace ainsi que la Moselle ne figurent pas dans la
base de données. Les DOM-TOM ont été retirés de l'analyse.
}
\examples{
\donttest{
library(tidyverse)

Evol <- data.frame()
Detail <- data.frame()
for (i in c(2018:2016)) {
  res  <- ValeurVenale(i)
  Evol <- rbind(Evol, as.data.frame(res$tab))
  Detail <- rbind(Detail, as.data.frame(res$Detail))
}

textes <- Evol \%>\% filter(Année == 2018)
ggplot(Evol, aes(x=Nature, y=Prix, color=factor(Année))) +
  geom_point() + geom_text(data=textes, aes(label=round(Prix,0)), size=4, hjust=1.2) +
  theme_bw() + labs(color="")

ggplot(Detail, aes(x=Nature, y=Prix, fill=factor(Année), color=factor(Année))) +
geom_point(position = position_jitterdodge()) +
  geom_boxplot(alpha=0.4) +
  geom_text(data=textes, aes(label=round(Prix,0)), size=3, hjust=-1.8) +
  theme_bw() + labs(fill="", color="", y="Prix (euro/ha)")
}
}
\author{
Bruciamacchie Max
}
