% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIfnVonf.R
\name{Vifn_Vonf}
\alias{Vifn_Vonf}
\title{Ratio entre volume EMERGE et volume IFN}
\format{
Renvoie le ratio sous forme d'objet sf ou de carte.
}
\source{
ONF
}
\usage{
Vifn_Vonf(ess = "09")
}
\arguments{
\item{ess}{= essence selon la codification de l'IFN}
}
\description{
La fonction Vifn_Vonf cartographie par sylvoécorégion le ratio entre le volume
EMERGE et le volume géométrique bois fort tige de l'IFN.
}
\examples{
df <- Tarif_EMERGE(df, df$Haut, df$C130)

}
\author{
Martin Marie-Laure
}
