% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voisins.R
\name{Voisins}
\alias{Voisins}
\title{Recherche voisins}
\usage{
Voisins(shp)
}
\arguments{
\item{shp}{couche géoréférencée contenant des POLYGON ou MULTIPOLYGON.}
}
\value{
La fonction renvoie l'objet sf fourni en entrée avec deux géométries supplémentaires,
l'une correspondant pour chaque polygone à l'union des voisins (geom2),
l'autre à l'union du polygone avec ses voisins (geom1).
}
\description{
Recherche de tous les polygones voisins d'un polygone.
}
\examples{
nc = st_read(system.file("shape/nc.shp", package="sf"))
shp <- voisins(nc)

par(mfrow=c(1,2))
st_geometry(shp) <- "geom1"
plot(st_geometry(shp[100,]),border="blue", lwd=4)
plot(st_geometry(nc[100,]), col="grey90", add=T)
plot(st_geometry(nc), border="red", lwd=1.5, add=T)

st_geometry(shp) <- "geom2"
plot(st_geometry(shp[100,]),border="blue", lwd=4)
plot(st_geometry(nc[100,]), col="grey90", add=T)
plot(st_geometry(nc), border="red", lwd=1.5, add=T)
par(mfrow=c(1,1))

}
\author{
Max Bruciamacchie
}
