% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zone Exploitable.R
\name{Zexpl}
\alias{Zexpl}
\title{Zones exploitables}
\usage{
Zexpl(r, shp)
}
\arguments{
\item{r}{= un modèle numérique de terrain}

\item{shp}{= la route au format sf}
}
\value{
La fonction renvoit le MNT
}
\description{
Affichage des zones exploitables (pente<0.30) par tracteur sur le MNT
}
\author{
camille, mathilde et nathan
}
