% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploiTracteur.R, R/Zone Exploitable.R
\name{ExploiTracteur}
\alias{ExploiTracteur}
\title{Zones exploitables}
\usage{
ExploiTracteur(r, shp)

ExploiTracteur(r, shp)
}
\arguments{
\item{r}{= un modèle numérique de terrain}

\item{shp}{= la route au format sf}

\item{amont}{= distance selon la pente exploitable en amont. Par défaut amont = 50m.}

\item{aval}{= distance selon la pente exploitable en aval. Par défaut aval = 150m.}

\item{pas}{= résolution.  Par défaut pas = 25m}

\item{conc}{= indice relatif de concavité. Plus cet indice est faible plus le contour sera découpé.}

\item{plane}{= valeur de pente en dessous de laquelle la pente est considérée comme plane.}
}
\value{
La fonction renvoit le MNT

La fonction renvoit le MNT
}
\description{
Affichage des zones exploitables (pente<0.30) par tracteur sur le MNT

Affichage des zones exploitables (pente<0.30) par tracteur sur le MNT
}
\author{
Max Bruciamacchie

Max Bruciamacchie
}
