% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteMnt.R
\name{SiteMnt}
\alias{SiteMnt}
\title{Modèle numérique de terrain}
\usage{
SiteMnt(shp, buffer = 50, zoom = 14, epsg = NULL)
}
\arguments{
\item{shp}{= périmètre de la zone au format sf.}

\item{buffer}{= taille du buffer. Par défaut buffer = 50 m.}

\item{zoom}{= un entier compris entre 1 et 14. Plus le zoom est élevé plus la résolution sera faible.}
}
\value{
La fonction renvoie un MNT dont la zone d'extension correspond à celle du polygone fourni en entrée.
}
\description{
Modèle numérique de terrain à partir du serveur AWS.
}
\examples{
\donttest{
library(elevatr)
library(raster)
library(sf)
perim <- FD \%>\% filter(IIDTN_FRT == "F10451Y")
r <- SiteMnt(perim)
par(mar = c(0,0,0,0))
plot(r, axes=F, box=F)
plot(st_geometry(perim), add=T)
}

}
\author{
Bruciamacchie Max
}
